requirements.txt
discord.py
requests
scikit-learn
joblib
python-dotenv